requirements.txt

Django==2.2.13
gunicorn==20.0.4
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
whitenoise==5.0.1
python-decouple==3.3      
psycopg2-binary==2.8.3
dj-database-url==0.5.0    
django-ckeditor==5.9.0
django-rest-framework==0.1.0    
django-cors-headers==3.2.1 
Pillow==7.1.2
django-cloudinary-storage==0.2.3   
cloudinary==1.20.0    